Change type body mass to f64
This simplifies interaction with other game code.

+++ source/rust/vndf/src/shared/game.rs
@@ -11,7 +11,7 @@ pub struct Body {
     pub position: Vec2<f64>,
     pub velocity: Vec2<f64>,
     pub force   : Vec2<f64>,
-    pub mass    : f32,
+    pub mass    : f64,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, RustcDecodable, RustcEncodable)]
